Reorganize exceptions and add middleware for exception handling
