+++ backend/services/__init__.py
@@ -1,5 +1,6 @@
 from .user import UserService
-from .permission import PermissionService, UserPermissionException
+from .permission import PermissionService
 from .role import RoleService
 from .github import GitHubService
 from .organization import OrganizationService
+from .exceptions import ResourceNotFoundException, UserPermissionException
