+++ backend/services/organization.py
@@ -6,7 +6,8 @@ from ..database import db_session
 from ..models.organization import Organization
 from ..entities.organization_entity import OrganizationEntity
 from ..models import User
-from .permission import PermissionService, UserPermissionException
+from .permission import PermissionService
+from .exceptions import UserPermissionException
 
 __authors__ = ["Ajay Gandecha", "Jade Keegan", "Brianna Ta", "Audrey Toney"]
 __copyright__ = "Copyright 2023"
