+++ backend/services/permission.py
@@ -10,18 +10,11 @@ from sqlalchemy.orm import Session
 from ..database import db_session
 from ..models import User, Permission, Role, RoleDetails
 from ..entities import UserEntity, PermissionEntity, RoleEntity
+from ..services.exceptions import UserPermissionException
 
 __authors__ = ["Kris Jordan"]
 __copyright__ = "Copyright 2023"
 __license__ = "MIT"
-
-
-class UserPermissionException(Exception):
-    """UserPermissionException is raised when a user attempts to perform an action they are not authorized to perform."""
-
-    def __init__(self, action: str, resource: str):
-        super().__init__(
-            f'Not authorized to perform `{action}` on `{resource}`')
 
 
 class PermissionService:
